pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "lux-video-gen"
version = "0.1.0"
description = "Lux: Advanced AI Video Generation Model - A Seedance 2.0 Competitor"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Lux Team"},
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "transformers>=4.40.0",
    "diffusers>=0.28.0",
    "accelerate>=0.30.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.23.0",
    "einops>=0.7.0",
    "rotary-embedding-torch>=0.6.0",
    "xformers>=0.0.25",
    "flash-attn>=2.5.0",
    "gradio>=4.20.0",
    "opencv-python>=4.9.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.4.9",
    "av>=12.0.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "tensorboard>=2.16.0",
    "matplotlib>=3.8.0",
    "decord>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
lux-train = "scripts.train:main"
lux-generate = "scripts.generate:main"
lux-ui = "lux_video_gen.ui.gradio_app:main"

[tool.setuptools.packages.find]
include = ["lux_video_gen*"]
